Split client capabilities by `ClientRoom` join preemption.
A client is needed which doesn't have preemption
in order to bootstrap `ClientRooms` with a list of
joined rooms.

Ideally we'd have a distinct capability that didn't rely
on the `allClient` just for joined rooms.

Then the ClientRooms factory would use an instance interner
the same way as the manager factory.

Then we'd remove the code that preloads joined rooms from
ClientRooms to avoid a bootstrapping problem.
That should instead be done by a listener the same way as
the ClientsInRoomsMap is updated, possibly at the same time.

The ClientsInRoomsMap will be the factory.
You can make sure to tie in the room state manager factory's
callback before creating the first ClientRooms.